Refactor health functions
