+++ src/lib.rs
@@ -1,15 +1,17 @@
 pub mod auth;
 pub mod dashboard_server;
 pub mod dispatcher;
+pub mod health;
 pub mod protocol;
 pub mod utils;
 
 pub use auth::UserRegistry;
 pub use dashboard_server::DashboardServer;
 pub use dispatcher::{
-    AppState, LogBuffer, LogBufferWriter, health_handler, model_handler, models_handler,
-    proxy_handler, run_worker, tags_handler,
+    AppState, LogBuffer, LogBufferWriter, model_handler, models_handler, proxy_handler, run_worker,
+    tags_handler,
 };
+pub use health::health_handler;
 pub use protocol::{
     BackendSnapshot, DashboardCmd, DashboardSnapshot, consumed_len, decode, encode,
 };
